Source code cleaned up
resolves #153
