--- fantasylol/schemas/game_state.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class GameState(str, Enum):
-    COMPLETED = "completed"
-    INPROGRESS = "inProgress"
-    UNSTARTED = "unstarted"
-    UNNEEDED = "unneeded"
