--- fantasylol/schemas/player_role.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class PlayerRole(str, Enum):
-    TOP = "top"
-    JUNGLE = "jungle"
-    MID = "mid"
-    BOTTOM = "bottom"
-    SUPPORT = "support"
-    NONE = "none"
