--- fantasylol/schemas/tournament_status.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class TournamentStatus(str, Enum):
-    COMPLETED = "completed"
-    ACTIVE = "active"
-    UPCOMING = "upcoming"
